[ncl] Cleanup React Native Core screen in native-component-list
