+++ apps/native-component-list/src/navigation/ReactNativeCoreStackNavigator.ts
@@ -1,6 +1,6 @@
 import BasicMaskExample from '../screens/BasicMaskScreen';
 import GLMaskExample from '../screens/MaskGLScreen';
-import ReactNativeCore from '../screens/ReactNativeCoreScreen';
+import ReactNativeCore from '../screens/ReactNativeCore/ReactNativeCoreScreen';
 import createStackNavigator from './createStackNavigator';
 import StackConfig from './StackConfig';
