+++ apps/native-component-list/src/screens/ReactNativeCore/ReactNativeCoreScreen.ios.tsx
@@ -1,32 +1,37 @@
 import React from 'react';
 import {
+  ActionSheetIOS,
   ActivityIndicator,
   Alert,
   Image,
   Picker,
+  ProgressViewIOS,
   RefreshControl,
+  SectionList,
+  SegmentedControlIOS,
   Slider,
   Switch,
   StatusBar,
-  SectionList,
   ScrollView,
   StyleSheet,
   Text,
   TextInput,
   TouchableOpacity,
   TouchableHighlight,
-  TouchableOpacityProps,
   View,
-  WebView,
+  TouchableOpacityProps,
 } from 'react-native';
+// @ts-ignore
+import TouchableBounce from 'react-native/Libraries/Components/Touchable/TouchableBounce';
+// @ts-ignore
+import { NavigationScreenProps, ScrollView as NavigationScrollView } from 'react-navigation';
+import WebView from 'react-native-webview';
 
-import Colors from '../constants/Colors';
-import Layout from '../constants/Layout';
-import { NavigationScreenProps } from 'react-navigation';
+import Colors from '../../constants/Colors';
+import Layout from '../../constants/Layout';
+import ModalExample from '../ModalExample';
 
 export default class ReactNativeCoreScreen extends React.Component<NavigationScreenProps> {
-  static path = '';
-
   static navigationOptions = {
     title: 'React Native Core',
   };
@@ -35,12 +40,39 @@ export default class ReactNativeCoreScreen extends React.Component<NavigationScr
     isRefreshing: false,
   };
 
-  _listView?: React.Component;
+  sections: Array<{ title: string, data: Array<() => JSX.Element> }>;
+
+  _sectionList?: React.Component;
+
+  constructor(props: any) {
+    super(props);
+
+    this.sections = [
+      { title: 'Vertical ScrollView, RefreshControl', data: [this._renderRefreshControl] },
+      { title: 'ActionSheetIOS', data: [this._renderActionSheet] },
+      { title: 'ActivityIndicator', data: [this._renderActivityIndicator] },
+      { title: 'Alert', data: [this._renderAlert] },
+      { title: 'Horizontal ScrollView', data: [this._renderHorizontalScrollView] },
+      { title: 'MaskView', data: [this._renderMaskView] },
+      { title: 'Modal', data: [this._renderModal] },
+      { title: 'Picker', data: [this._renderPicker] },
+      { title: 'ProgressView', data: [this._renderProgressView] },
+      { title: 'SegmentedControl', data: [this._renderSegmentedControl] },
+      { title: 'Slider', data: [this._renderSlider] },
+      { title: 'StatusBar', data: [this._renderStatusBar] },
+      { title: 'Switch', data: [this._renderSwitch] },
+      { title: 'Text', data: [this._renderText] },
+      { title: 'TextInput', data: [this._renderTextInput] },
+      { title: 'Touchables', data: [this._renderTouchables] },
+      // { title: 'View', data: [this._renderView] },
+      { title: 'WebView', data: [this._renderWebView] },
+    ];
+  }
 
   render() {
     return (
       <SectionList
-        ref={view => (this._listView = view!)}
+        ref={view => (this._sectionList = view!)}
         keyboardShouldPersistTaps="handled"
         keyboardDismissMode="on-drag"
         refreshControl={
@@ -49,28 +81,13 @@ export default class ReactNativeCoreScreen extends React.Component<NavigationScr
         keyExtractor={(item, index) => `${index}`}
         removeClippedSubviews={false}
         contentContainerStyle={{ backgroundColor: '#fff' }}
-        sections={[
-          { title: 'Vertical ScrollView, RefreshControl', data: [this._renderRefreshControl] },
-          { title: 'ActivityIndicator', data: [this._renderActivityIndicator] },
-          { title: 'Alert', data: [this._renderAlert] },
-          { title: 'Horizontal ScrollView', data: [this._renderHorizontalScrollView] },
-          { title: 'MaskView', data: [this._renderMaskView] },
-          { title: 'Picker', data: [this._renderPicker] },
-          { title: 'Slider', data: [this._renderSlider] },
-          { title: 'StatusBar', data: [this._renderStatusBar] },
-          { title: 'Switch', data: [this._renderSwitch] },
-          { title: 'Text', data: [this._renderText] },
-          { title: 'TextInput', data: [this._renderTextInput] },
-          { title: 'Touchables', data: [this._renderTouchables] },
-          { title: 'WebView', data: [this._renderWebView] },
-        ]}
+        sections={this.sections}
         renderItem={this._renderItem}
+        renderScrollComponent={props => <NavigationScrollView {...props} />}
         renderSectionHeader={this._renderSectionHeader}
       />
     );
   }
-
-  _renderItem = ({ item }: { item: () => JSX.Element }) => item();
 
   _onRefresh = () => {
     this.setState({ isRefreshing: true });
@@ -80,8 +97,7 @@ export default class ReactNativeCoreScreen extends React.Component<NavigationScr
   }
 
   _scrollToTop = () => {
-    // @ts-ignore
-    this._listView!.scrollTo({ x: 0, y: 0 });
+    this._sectionList!.scrollTo({ x: 0, y: 0 });
   }
 
   _renderMaskView = () => {
@@ -97,14 +113,59 @@ export default class ReactNativeCoreScreen extends React.Component<NavigationScr
     );
   }
 
-  _renderRefreshControl = () => (
-    <View style={{ padding: 10 }}>
-      <Text>
-        This screen is a vertical ScrollView, try the pull to refresh gesture to see the
-        RefreshControl.
-      </Text>
-    </View>
-  )
+  _renderModal = () => {
+    return <ModalExample />;
+  }
+
+  _renderRefreshControl = () => {
+    return (
+      <View style={{ padding: 10 }}>
+        <Text>
+          This screen is a vertical ScrollView, try the pull to refresh gesture to see the
+          RefreshControl.
+        </Text>
+      </View>
+    );
+  }
+
+  _renderActionSheet = () => {
+    const showActionSheet = () => {
+      ActionSheetIOS.showActionSheetWithOptions(
+        {
+          options: ['Option 0', 'Option 1', 'Delete', 'Cancel'],
+          cancelButtonIndex: 3,
+          destructiveButtonIndex: 2,
+        },
+        buttonIndex => {
+          console.log({ buttonIndex });
+        }
+      );
+    };
+
+    const showShareSheet = () => {
+      ActionSheetIOS.showShareActionSheetWithOptions(
+        {
+          url: 'https://expo.io',
+          message: 'message to go with the shared url',
+          subject: 'a subject to go in the email heading',
+        },
+        error => alert(error),
+        (success, method) => {
+          if (success) {
+            alert(`Shared via ${method}`);
+          }
+        }
+      );
+    };
+
+    return (
+      <View style={{ flexDirection: 'row', padding: 10 }}>
+        <Button onPress={showActionSheet}>Action sheet</Button>
+
+        <Button onPress={showShareSheet}>Share sheet</Button>
+      </View>
+    );
+  }
 
   _renderActivityIndicator = () => {
     const Spacer = () => <View style={{ marginRight: 10 }} />;
@@ -126,7 +187,9 @@ export default class ReactNativeCoreScreen extends React.Component<NavigationScr
   }
 
   _renderAlert = () => {
-    const showPrompt = () => {};
+    const showPrompt = () => {
+      Alert.prompt('Enter a value', undefined, text => console.log(`You entered ${text}`));
+    };
 
     const showAlert = () => {
       Alert.alert('Alert Title', 'My Alert Msg', [
@@ -168,17 +231,17 @@ export default class ReactNativeCoreScreen extends React.Component<NavigationScr
     return (
       <ScrollView pagingEnabled directionalLockEnabled horizontal>
         <Image
-          source={require('../../assets/images/example1.jpg')}
+          source={require('../../../assets/images/example1.jpg')}
           style={imageStyle}
           resizeMode="cover"
         />
         <Image
-          source={require('../../assets/images/example2.jpg')}
+          source={require('../../../assets/images/example2.jpg')}
           style={imageStyle}
           resizeMode="cover"
         />
         <Image
-          source={require('../../assets/images/example3.jpg')}
+          source={require('../../../assets/images/example3.jpg')}
           style={imageStyle}
           resizeMode="cover"
         />
@@ -188,6 +251,21 @@ export default class ReactNativeCoreScreen extends React.Component<NavigationScr
 
   _renderPicker = () => {
     return <PickerExample />;
+  }
+
+  _renderProgressView = () => {
+    return (
+      <View style={{ padding: 10, paddingBottom: 30 }}>
+        <ProgressViewExample initialProgress={0} />
+        <ProgressViewExample progressTintColor="red" initialProgress={0.4} />
+        <ProgressViewExample progressTintColor="orange" initialProgress={0.6} />
+        <ProgressViewExample progressTintColor="yellow" initialProgress={0.8} />
+      </View>
+    );
+  }
+
+  _renderSegmentedControl = () => {
+    return <SegmentedControlExample />;
   }
 
   _renderSlider = () => {
@@ -273,9 +351,21 @@ export default class ReactNativeCoreScreen extends React.Component<NavigationScr
         <TouchableOpacity style={buttonStyle} onPress={() => {}}>
           <Text style={buttonText}>Opacity!</Text>
         </TouchableOpacity>
+
+        <TouchableBounce style={buttonStyle} onPress={() => {}}>
+          <Text style={buttonText}>Bounce!</Text>
+        </TouchableBounce>
       </View>
     );
   }
+
+  // _renderView = () => {
+  //   // Don't know what to put here
+  //   return (
+  //     <View>
+  //     </View>
+  //   );
+  // }
 
   _renderWebView = () => {
     return (
@@ -298,6 +388,8 @@ export default class ReactNativeCoreScreen extends React.Component<NavigationScr
     );
   }
 
+  _renderItem = ({ item }: { item: () => JSX.Element }) => item();
+
   _renderSectionHeader = ({ section }: any) => (
     <View style={styles.sectionHeader}>
       <Text>{section.title}</Text>
@@ -312,7 +404,9 @@ class PickerExample extends React.Component {
 
   render() {
     return (
+      // @ts-ignore
       <Picker
+        st
         selectedValue={this.state.language}
         onValueChange={lang => this.setState({ language: lang })}
       >
@@ -321,6 +415,76 @@ class PickerExample extends React.Component {
         <Picker.Item label="Objective C" value="objc" />
         <Picker.Item label="Swift" value="swift" />
       </Picker>
+    );
+  }
+}
+
+interface ProgressViewExampleProps {
+  progressTintColor?: string;
+  initialProgress: number;
+}
+
+interface ProgressViewExampleState {
+  progress: number;
+}
+
+class ProgressViewExample extends React.Component<ProgressViewExampleProps, ProgressViewExampleState> {
+  constructor(props: ProgressViewExampleProps) {
+    super(props);
+
+    this.state = {
+      progress: props.initialProgress,
+    };
+  }
+
+  render() {
+    const progressStyle = { marginTop: 20 };
+
+    return (
+      <ProgressViewIOS
+        style={progressStyle}
+        progressTintColor={this.props.progressTintColor}
+        progress={this.state.progress}
+      />
+    );
+  }
+}
+
+class SegmentedControlExample extends React.Component {
+  state = {
+    selectedIndex: 0,
+  };
+
+  render() {
+    let tintColor;
+    switch (this.state.selectedIndex) {
+      case 0:
+        tintColor = 'black';
+        break;
+      case 1:
+        tintColor = Colors.tintColor;
+        break;
+      case 2:
+        tintColor = 'green';
+        break;
+      case 3:
+        tintColor = 'purple';
+        break;
+    }
+
+    return (
+      <View style={{ margin: 10 }}>
+        <SegmentedControlIOS
+          values={['One', 'Two', 'Three', 'Four']}
+          tintColor={tintColor}
+          selectedIndex={this.state.selectedIndex}
+          onChange={event => {
+            this.setState({
+              selectedIndex: event.nativeEvent.selectedSegmentIndex,
+            });
+          }}
+        />
+      </View>
     );
   }
 }
@@ -354,8 +518,13 @@ class SliderExample extends React.Component<SliderExampleProps, SliderExampleSta
 
     return (
       <View>
-        <Text style={textStyle}>Value: {this.state.value && +this.state.value.toFixed(3)}</Text>
+        <View style={{ padding: 10 }}>
+          <Text style={textStyle}>Value: {this.state.value && +this.state.value.toFixed(3)}</Text>
+        </View>
+
         <Slider {...this.props} onValueChange={value => this.setState({ value })} />
+
+        <View style={{ marginBottom: 10 }} />
       </View>
     );
   }
@@ -426,11 +595,13 @@ class TextInputExample extends React.Component {
   }
 }
 
-const Button: React.FunctionComponent<TouchableOpacityProps> = props => (
-  <TouchableOpacity onPress={props.onPress} style={styles.button}>
-    <Text style={styles.buttonText}>{props.children}</Text>
-  </TouchableOpacity>
-);
+const Button: React.FunctionComponent<TouchableOpacityProps> = (props) => {
+  return (
+    <TouchableOpacity onPress={props.onPress} style={styles.button}>
+      <Text style={styles.buttonText}>{props.children}</Text>
+    </TouchableOpacity>
+  );
+};
 
 const styles = StyleSheet.create({
   container: {
