+++ apps/native-component-list/src/screens/ReactNativeCore/ReactNativeCoreScreen.tsx
